5.49
Trying to fix -> https://github.com/tonikelope/megabasterd/issues/32

+++ src/main/java/com/tonikelope/megabasterd/ProgressMeter.java
@@ -63,7 +63,7 @@ public final class ProgressMeter implements Runnable, SecureSingleThreadNotifiab
         _progress = _transference.getProgress();
 
         while (!_exit || !_transference.getPartialProgress().isEmpty()) {
-            Integer reads;
+            Long reads;
 
             while ((reads = _transference.getPartialProgress().poll()) != null) {
                 _progress += reads;
